Add rmdir
Currently ( mkdir -p a/b && rm -r a ) is broken. It looks like rm is doing
opendir, rmdir, then readdir, and it gets angry when b is yielded from readdir.
b is actually removed, but rm aborts before removing a.

According to POSIX for readdir:
    If a file is removed from or added to the directory after the most recent
    call to opendir() or rewinddir(), whether a subsequent call to readdir()
    returns an entry for that file is unspecified.

Still, it seems like taking a snapshot on opendir is not OK.